router and index updated
